puppet-refactor-v1/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
PyYAML>=6.0.1
qdrant-client>=1.7.0
